requirements.txt
discord.py
bs4
requests
groq
pillow
python-dotenv
colorlog